Allow underlying element props to be passed to typography components
This means we can treat typography components the same as the underlying
html element (e.g set an id). Maybe we want to be more restrictive, but
it feels okay to allow all props that for example a <h1> tag is
expecting.

To do this we had to use props spreading, which has been disabled for
each file. Again I think this is okay so we don't have to reinvent the
expected props for these elements

[#181325681]